Added variables for Assignment 1 Finished Assignment 1 - you can double check and we can see if we can add more comments
Started assignment 2 - I'm pretty sure that the code is right but I don't know how to check it properly

Hey, I did more assignment 2 . For assignment 1, I think we did things differently with how we added the commision as cost. I got a negative because the commission from both buying and selling was added to the initial buying of stocks which is higher than the price she sold it.
Do you have an email? I lost your email and I think we can communicate better there

Exercise 3 working again